+++ functions/lib/index.js
@@ -1351,13 +1351,12 @@ async function postToBuffer(questionText, isMorning, totalUsers) {
                     query: `
             mutation CreatePost($input: CreatePostInput!) {
               createPost(input: $input) {
-                ... on PostActionPayload {
+                ... on PostActionSuccess {
                   post {
                     id
-                    status
                   }
                 }
-                ... on Error {
+                ... on PostActionError {
                   message
                 }
               }